[encode] Allow downloading YouTube videos
This commit adds functionality to download YouTube videos. It uses a
pattern match in case of a stream to ensure the stream is coming from
YouTube. This is done to don't disrupt other stream methods.

Furthermore ffmpeg input and output arguments are getting split since we
need to differentiate between them, if we pass them to youtube dlp. If
we are handling a normal file they just get merged together.