[blacklist-extensions.lua] Check if file without extension is directory before removing it from playlist
When mpv uses `directory-mode=auto` or `directory-mode=lazy` (the default in mpv 0.37.0 it seems) it adds subdirectories themselves to the playlist instead of their contents, only to expand them when they are played. When `remove_files_without_extension` is set to `true`, directories which do not have a dot in their name are also removed from the playlist. I added a check to prevent this.